src/requirements.txt
fastapi[all]
python-dotenv
uuid
sqlalchemy